requirements.txt
Flask==0.10.1
Flask-RESTful==0.3.5
python-dotenv==0.4.0
pytest==2.9.1
